Updates.
git-svn-id: http://warpproject.org/svn/WARP@3835 dd934ca3-7c2e-414f-9084-136bfdd87501

+++ ReferenceDesigns/w3_802.11/c/wlan_mac_high_framework/include/wlan_mac_dl_list.h
@@ -34,7 +34,7 @@ struct dl_entry{
 typedef struct {
 	dl_entry* first;
 	dl_entry* last;
-	u16 length;
+	u32       length;
 } dl_list;
